+++ src/parser/parser3.c
@@ -6,7 +6,7 @@
 /*   By: jcat <joaoteix@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/01 01:53:49 by jcat              #+#    #+#             */
-/*   Updated: 2024/05/05 16:40:20 by joaoteix         ###   ########.fr       */
+/*   Updated: 2024/05/07 19:55:43 by joaoteix         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -109,7 +109,7 @@ bool	parse_vec3(const char *str, t_vec3 *v)
 bool	parse_rgb(const char *str, t_color3 *color)
 {
 	char const		**strs = (char const **)ft_split(str, ',');
-	double *const	comp[] = {&color->r, &color->g, &color->b};
+	float *const	comp[] = {&color->r, &color->g, &color->b};
 	int				i;
 	int				tmp;
 	bool			valid;
@@ -126,7 +126,7 @@ bool	parse_rgb(const char *str, t_color3 *color)
 			valid = false;
 			break ;
 		}
-		*(comp[i]) = (double)tmp / 255.f;
+		*(comp[i]) = (float)tmp / 255.f;
 		++i;
 	}
 	if (i < 2 || strs[i] != NULL)
